refactor: actionTypes를 actions파일 내부로 이동
