+++ src/App.js
@@ -8,20 +8,18 @@ import { PATH } from 'constants/path';
 
 function App() {
   return (
-    <>
-      <Suspense fallback={<ImgWrapper src={spinner} />}>
-        <BrowserRouter>
-          <Routes>
-            <Route element={<Layout />}>
-              <Route path="/" exact element={<ProductList />} />
-              <Route path={`${PATH.PRODUCT}/:id`} element={<Product />} />
-              <Route path={PATH.CART} exact element={<Cart />} />
-              <Route path="*" element={<NotFound />} />
-            </Route>
-          </Routes>
-        </BrowserRouter>
-      </Suspense>
-    </>
+    <Suspense fallback={<ImgWrapper src={spinner} />}>
+      <BrowserRouter>
+        <Routes>
+          <Route element={<Layout />}>
+            <Route path="/" exact element={<ProductList />} />
+            <Route path={`${PATH.PRODUCT}/:id`} element={<Product />} />
+            <Route path={PATH.CART} exact element={<Cart />} />
+            <Route path="*" element={<NotFound />} />
+          </Route>
+        </Routes>
+      </BrowserRouter>
+    </Suspense>
   );
 }
