--- src/reducers/cart/cart.actionTypes.js
@@ -1,6 +1,0 @@
-export default {
-  GET_CART: 'GET_CART',
-  GET_CART_SUCCESS: 'GET_CART_SUCCESS',
-  GET_CART_ERROR: 'GET_CART_ERROR',
-  ADD_CART_ITEM: 'ADD_CART_ITEM',
-};
