+++ src/reducers/cart/cart.actions.js
@@ -1,4 +1,9 @@
-import actionTypes from 'reducers/cart/cart.actionTypes';
+export const actionTypes = {
+  GET_CART: 'GET_CART',
+  GET_CART_SUCCESS: 'GET_CART_SUCCESS',
+  GET_CART_ERROR: 'GET_CART_ERROR',
+  ADD_CART_ITEM: 'ADD_CART_ITEM',
+};
 
 export const getCart = () => ({
   type: actionTypes.GET_CART,
