--- src/reducers/product/product.actionTypes.js
@@ -1,5 +1,0 @@
-export default {
-  GET_PRODUCT: 'GET_PRODUCT',
-  GET_PRODUCT_SUCCESS: 'GET_PRODUCT_SUCCESS',
-  GET_PRODUCT_ERROR: 'GET_PRODUCT_ERROR',
-};
