--- src/reducers/products/products.actionTypes.js
@@ -1,5 +1,0 @@
-export default {
-  GET_PRODUCTS: 'GET_PRODUCTS',
-  GET_PRODUCTS_SUCCESS: 'GET_PRODUCTS_SUCCESS',
-  GET_PRODUCTS_ERROR: 'GET_PRODUCTS_ERROR',
-};
