+++ src/reducers/products/products.actions.js
@@ -1,4 +1,8 @@
-import actionTypes from 'reducers/products/products.actionTypes';
+export const actionTypes = {
+  GET_PRODUCTS: 'GET_PRODUCTS',
+  GET_PRODUCTS_SUCCESS: 'GET_PRODUCTS_SUCCESS',
+  GET_PRODUCTS_ERROR: 'GET_PRODUCTS_ERROR',
+};
 
 export const getProducts = () => ({
   type: actionTypes.GET_PRODUCTS,
